`@remotion/studio`: Keep caption editing inline
